Enable querying meals.

+++ src/main/java/data/create-database.sql
@@ -94,7 +94,7 @@ CREATE TABLE meal_items
     id int PRIMARY KEY AUTO_INCREMENT,
     meal_id INT,
     food_id INT NOT NULL,
-    quantity INT NOT NULL,
+    quantity FLOAT NOT NULL,
     FOREIGN KEY (meal_id) REFERENCES meals(id)
     -- TO DO: disabling foreign key reference temporarily while using mock food data
     -- until we finalize foods in database and move off of using mocks entirely
